Implement ability to execute command in Bridge environment shell
Currently, any arguments passed into the connect command following '--' are stored in the `CommandExecuteArguments` property of the `CommandLineArgumentsManager`. Until now, this property was ignored.
Now `CommandExecuteArguments` are passed into `LaunchTerminalWithEnv()` whenever `performLaunch` is true, which then executes the provided command in the launched environment.

To avoid changing the way the shell is launched in each platform, the Windows method appends the command to the output script and the bash method executes the specified command instead of bash.